requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# PyTorch - Versiones compatibles actuales
torch>=2.5.0,<3.0.0
torchvision>=0.20.0,<1.0.0

# CLIP
git+https://github.com/openai/CLIP.git

# Vector search
faiss-cpu>=1.7.3,<2.0.0

# Image processing
pillow>=10.0.0,<11.0.0
numpy>=1.24.0,<2.0.0
opencv-python-headless>=4.8.0,<5.0.0

# Core utilities
pydantic[email]>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-dateutil>=2.8.0,<3.0.0
tqdm>=4.65.0,<5.0.0
httpx>=0.25.0,<1.0.0

# Development (optional)
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.0,<1.0.0